requirements.txt
Django==5.2.5
django-environ==0.12.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
channels==4.1.0
daphne==4.1.2
channels-redis==4.2.0
redis==4.6.0
asgiref==3.9.1
pyjwt
drf-spectacular
requests